Fix available workloads returned by resolver
Platform checks should take restrictions on base workloads into account,
as clarified in https://github.com/dotnet/designs/pull/230.
This means that the CLI can no longer trivially display the RIDs for
the platforms on which the workloads are available. This could be added
back in future using the RID graph to coalesce values, but would probably
need additional work to ensure the values were user friendly.

The resolver no longer returns workload definitions. The ensures clients
use the resolver's APIs that work on the composed manifest data. Having
all code that deals with manifest data direct in a central location
helps keep behaviour consistent.

It also does not be return abstract or empty workloads as these
cannot be installed.